# src/main/resources/static/api/openapi.yml
openapi: 3.1.0
info:
  title: BOS API documentation
  description: |
    내꿈은건물주  API 문서

  version: 0.0.1-SNAPSHOT
  contact: {}
...
paths:
...
  /push/test/expo:
    post:
      tags:
        - push
      operationId: sendExpoTestPush
      summary: "Expo 푸시 알림 테스트"
      description: |-
        Expo Push Notification API를 사용하여 특정 사용자에게 테스트 푸시 알림을 전송합니다.

        ## 기능
        - userId로 user_devices 테이블에서 토큰 조회
        - 조회된 모든 디바이스에 푸시 전송
        - Expo API 에러 발생 시 상세 정보 포함하여 응답

        ## 응답 정보
        - 각 디바이스별 전송 성공/실패 여부
        - 실패한 경우 해당 토큰 정보 (token, updatedAt 포함)
        - 전체 전송 통계 (성공/실패 건수)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - title
                - body
              properties:
                userId:
                  type: integer
                  format: int64
                  description: 푸시를 받을 사용자 ID
                  example: 1
                title:
                  type: string
                  description: 푸시 알림 제목
                  example: "테스트 알림"
                body:
                  type: string
                  description: 푸시 알림 본문
                  example: "테스트 메시지입니다"
      responses:
        "200":
          description: 푸시 전송 요청 완료 (성공 여부는 응답 본문 확인)
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                  - sentCount
                properties:
                  success:
                    type: boolean
                    description: 전체 전송 성공 여부 (1건 이상 성공 시 true)
                    example: true
                  message:
                    type: string
                    description: 전송 결과 메시지
                    example: "푸시 전송 성공: 2건"
                  sentCount:
                    type: integer
                    description: 성공적으로 전송된 건수
                    example: 2
                  devices:
                    type: array
                    description: 전송 대상 디바이스 목록
                    items:
                      type: object
                      required:
                        - userId
                        - token
                        - updatedAt
                        - success
                      properties:
                        userId:
                          type: integer
                          format: int64
                          description: 사용자 ID
                          example: 1
                        token:
                          type: string
                          description: Expo Push Token
                          example: "ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx]"
                        platform:
                          type: string
                          nullable: true
                          description: 플랫폼 (IOS, ANDROID)
                          example: "IOS"
                        updatedAt:
                          type: string
                          format: date-time
                          description: 토큰 마지막 업데이트 시간
                          example: "2024-01-15T10:30:00Z"
                        success:
                          type: boolean
                          description: 해당 디바이스 전송 성공 여부
                          example: true
                        ticketId:
                          type: string
                          nullable: true
                          description: Expo Push Ticket ID (전송 성공 시)
                          example: "abc123-def456-ghi789"
                        expoErrorMessage:
                          type: string
                          nullable: true
                          description: Expo 서버에서 반환한 에러 메시지
                          example: "DeviceNotRegistered"
                        expoErrorDetails:
                          type: object
                          nullable: true
                          description: Expo 서버에서 반환한 에러 상세 정보
                          additionalProperties: true
                  error:
                    type: object
                    nullable: true
                    description: 에러 발생 시 상세 정보
                    properties:
                      errorMessage:
                        type: string
                        description: 에러 메시지
                        example: "토큰 ExponentPushToken[xxx] 전송 실패"
                      failedDevices:
                        type: array
                        description: 전송 실패한 디바이스 목록
                        items:
                          type: object
                          properties:
                            userId:
                              type: integer
                              format: int64
                            token:
                              type: string
                            platform:
                              type: string
                              nullable: true
                            updatedAt:
                              type: string
                              format: date-time
                            success:
                              type: boolean
              examples:
                success:
                  summary: 전송 성공
                  value:
                    success: true
                    message: "푸시 전송 성공: 2건"
                    sentCount: 2
                    devices:
                      - userId: 1
                        token: "ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx1]"
                        platform: "IOS"
                        updatedAt: "2024-01-15T10:30:00Z"
                        success: true
                        ticketId: "abc123-def456-ghi789"
                        expoErrorMessage: null
                        expoErrorDetails: null
                      - userId: 1
                        token: "ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx2]"
                        platform: "ANDROID"
                        updatedAt: "2024-01-15T09:20:00Z"
                        success: true
                        ticketId: "xyz987-uvw654-rst321"
                        expoErrorMessage: null
                        expoErrorDetails: null
                    error: null
                partial_failure:
                  summary: 일부 전송 실패 (Expo API 에러 포함)
                  value:
                    success: true
                    message: "푸시 전송 완료: 성공 1건, 실패 1건"
                    sentCount: 1
                    devices:
                      - userId: 1
                        token: "ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx1]"
                        platform: "IOS"
                        updatedAt: "2024-01-15T10:30:00Z"
                        success: true
                        ticketId: "abc123-def456-ghi789"
                        expoErrorMessage: null
                        expoErrorDetails: null
                      - userId: 1
                        token: "ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx2]"
                        platform: "ANDROID"
                        updatedAt: "2024-01-10T08:00:00Z"
                        success: false
                        ticketId: null
                        expoErrorMessage: "DeviceNotRegistered"
                        expoErrorDetails:
                          error: "DeviceNotRegistered"
                          fault: "developer"
                    error:
                      errorMessage: "토큰 ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx2]: DeviceNotRegistered"
                      failedDevices:
                        - userId: 1
                          token: "ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx2]"
                          platform: "ANDROID"
                          updatedAt: "2024-01-10T08:00:00Z"
                          success: false
                          ticketId: null
                          expoErrorMessage: "DeviceNotRegistered"
                          expoErrorDetails:
                            error: "DeviceNotRegistered"
                            fault: "developer"
                no_devices:
                  summary: 등록된 디바이스 없음
                  value:
                    success: false
                    message: "등록된 디바이스가 없습니다. 먼저 디바이스를 등록해주세요."
                    sentCount: 0
                    devices: null
                    error: null
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorCode:
                    type: string
                    example: "INVALID_REQUEST"
                  message:
                    type: string
                    example: "요청값이 올바르지 않습니다."
        "500":
          $ref: "#/components/responses/InternalServerErrorResponseSpec"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    InternalServerErrorResponseSpec:
      description: 내부 오류가 생겼습니다. 해당 상황 발생 시, 관리자에게 이야기해주세요.
      content:
        application/json:
          schema:
            properties:
              statusCode:
                type: number
                example: 500
              message:
                type: string
                example: Internal Server Error
...